refactor(mobile): moving providers
