+++ mobile/src/kernel/navigation/WalletNavigator.tsx
@@ -1,12 +1,14 @@
 import {useTheme} from '@yoroi/theme'
 
-import {createStackNavigator} from '@react-navigation/stack'
+import {
+  StackNavigationOptions,
+  createStackNavigator,
+} from '@react-navigation/stack'
 import * as React from 'react'
 
 import {DashboardNavigator} from '~/features/Dashboard/DashboardNavigator'
 import {CatalystNavigator} from '~/features/RegisterCatalyst/CatalystNavigator'
 import {ReviewTxNavigator} from '~/features/ReviewTx/ReviewTxNavigator'
-import {SearchProvider} from '~/features/Search/SearchContext'
 import {SettingsScreenNavigator} from '~/features/Settings/ui/navigation/SettingsScreenNavigator'
 import {NetworkTag} from '~/features/Settings/ui/shared/NetworkTag'
 import {SetupWalletNavigator} from '~/features/SetupWallet/SetupWalletNavigator'
@@ -23,68 +25,69 @@ export const WalletNavigator = () => {
   const strings = useStrings()
   const {palette: p} = useTheme()
 
-  const navOptions = React.useMemo(() => defaultStackNavigationOptions(p), [p])
+  const navOptions = React.useMemo(
+    () => ({...defaultStackNavigationOptions(p), headerLeft: undefined}),
+    [p],
+  )
+
+  const stackOptions: StackNavigationOptions = React.useMemo(
+    () => ({
+      title: strings.global.walletSelectionScreenHeader,
+      headerTitle: ({children}) => (
+        <NetworkTag directChangeActive>{children}</NetworkTag>
+      ),
+    }),
+    [strings.global.walletSelectionScreenHeader],
+  )
 
   return (
-    <SearchProvider>
-      <Stack.Navigator
-        screenOptions={{
-          ...navOptions,
-          headerLeft: undefined,
-        }}
-      >
-        <Stack.Screen
-          name="wallet-selection"
-          options={{
-            title: strings.global.walletSelectionScreenHeader,
-            headerTitle: ({children}) => (
-              <NetworkTag directChangeActive>{children}</NetworkTag>
-            ),
-          }}
-          getComponent={() => SelectWalletFromList}
-        />
+    <Stack.Navigator screenOptions={navOptions}>
+      <Stack.Screen
+        name="wallet-selection"
+        options={stackOptions}
+        getComponent={() => SelectWalletFromList}
+      />
 
-        <Stack.Screen //
-          name="setup-wallet"
-          options={{headerShown: false}}
-          getComponent={() => SetupWalletNavigator}
-        />
+      <Stack.Screen //
+        name="setup-wallet"
+        options={{headerShown: false}}
+        getComponent={() => SetupWalletNavigator}
+      />
 
-        <Stack.Screen
-          name="main-wallet-routes"
-          options={{headerShown: false}}
-          getComponent={() => WalletTabNavigator}
-        />
+      <Stack.Screen
+        name="main-wallet-routes"
+        options={{headerShown: false}}
+        getComponent={() => WalletTabNavigator}
+      />
 
-        <Stack.Screen
-          name="review-tx-routes"
-          options={{headerShown: false}}
-          getComponent={() => ReviewTxNavigator}
-        />
+      <Stack.Screen
+        name="review-tx-routes"
+        options={{headerShown: false}}
+        getComponent={() => ReviewTxNavigator}
+      />
 
-        <Stack.Screen
-          name="settings"
-          options={{headerShown: false}}
-          getComponent={() => SettingsScreenNavigator}
-        />
+      <Stack.Screen
+        name="settings"
+        options={{headerShown: false}}
+        getComponent={() => SettingsScreenNavigator}
+      />
 
-        <Stack.Screen
-          name="staking-dashboard"
-          options={{headerShown: false}}
-          getComponent={() => DashboardNavigator}
-        />
+      <Stack.Screen
+        name="staking-dashboard"
+        options={{headerShown: false}}
+        getComponent={() => DashboardNavigator}
+      />
 
-        <Stack.Screen
-          name="governance"
-          options={{headerShown: false}}
-          getComponent={() => GovernanceNavigator}
-        />
-        <Stack.Screen
-          name="voting-registration"
-          options={{headerShown: false}}
-          getComponent={() => CatalystNavigator}
-        />
-      </Stack.Navigator>
-    </SearchProvider>
+      <Stack.Screen
+        name="governance"
+        options={{headerShown: false}}
+        getComponent={() => GovernanceNavigator}
+      />
+      <Stack.Screen
+        name="voting-registration"
+        options={{headerShown: false}}
+        getComponent={() => CatalystNavigator}
+      />
+    </Stack.Navigator>
   )
 }
